+++ doc/source/doxygen-docs/changelog.md
@@ -5,6 +5,8 @@
   - ptg-configurator:
     - Show selected PTG path output motion command.
 - Changes in libraries:
+  - \ref mrpt_containers_grp
+    - New methods mrpt::containers::bimap::erase_by_key(),mrpt::containers::bimap::erase_by_value()
   - \ref mrpt_system_grp
     - Backwards-compatible change: New function mrpt::system::InvalidTimeStamp() used now inside the macro INVALID_TIMESTAMP, so the macro always returns a const reference instead of returning by value.
 - BUG FIXES: